added hash function

+++ Marakulin/archiver/hashFunc.h
@@ -1,5 +1,5 @@
 #include "../include/logs.h"
 #include "compressFile.h"
 
-unsigned int HashFAQ6(const char * str, off_t length);
+off_t HashFAQ6(const char * str, off_t length);
 off_t getHashSum(int fd);
